requirements.txt
numpy==2.0.0
Pillow==9.0.0
Pillow==10.4.0
Requests==2.32.3
